tools/trending-flagger/trend-detector/requirements.txt
aiohttp
psycopg2-binary
redis
numpy
scikit-learn
pyyaml